Spectrogram rebuilds upon maximum note changes
Ensure that mMatrix is reassigned to the correct dimensions once the
maximum note changes, which affects sigma, decimation rate, number of
rows, and number of columns.

+++ src/SuperSpectrogramWindow.h
@@ -64,7 +64,7 @@ private:
 #endif
 
    size_t mDetailLevel = 7;
-   size_t mMaxFreq = 1;
+   double mMaxFreq = 1.0;
    size_t mNoiseFloor = -70;
 
    enum {
